# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Memory Atlas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/weave": {
      "post": {
        "summary": "Run Weave",
        "operationId": "run_weave_api_weave_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeaveResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/weave/{session_id}": {
      "get": {
        "summary": "Latest Weave",
        "operationId": "latest_weave_api_weave__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeaveResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/clarify": {
      "post": {
        "summary": "Clarify",
        "description": "At most two short follow-up questions before weaving (docs/ART_DIRECTION.md \u00a7D).",
        "operationId": "clarify_api_clarify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClarifyResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/generate-asset": {
      "post": {
        "summary": "Generate Asset",
        "operationId": "generate_asset_api_generate_asset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAssetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/art/{name}": {
      "get": {
        "summary": "Art File",
        "operationId": "art_file_api_art__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtDirectionOut": {
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/ArtRequestOut"
            },
            "type": "array",
            "title": "Requests"
          },
          "annotations": {
            "items": {
              "$ref": "#/components/schemas/AnnotationOut"
            },
            "type": "array",
            "title": "Annotations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requests",
          "annotations"
        ],
        "title": "ArtDirectionOut"
      },
...
      "ClarifyQuestion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source_id",
          "text"
        ],
        "title": "ClarifyQuestion"
      },
      "ClarifyResult": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/ClarifyQuestion"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Questions"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoSeen"
            },
            "type": "array",
            "title": "Photos",
            "default": []
          },
          "voices": {
            "items": {
              "$ref": "#/components/schemas/VoiceHeard"
            },
            "type": "array",
            "title": "Voices",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "questions"
        ],
        "title": "ClarifyResult",
        "description": "The 'understand' step: what each photo shows, what each voice note says, and \u2264N questions for what can't be seen."
      },
...
      "DuplicateGroup": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 2,
            "title": "Source Ids"
          },
          "kept_source_id": {
            "type": "string",
            "format": "uuid",
            "title": "Kept Source Id"
          },
          "reason": {
            "type": "string",
            "enum": [
              "exact",
              "perceptual",
              "embedding"
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source_ids",
          "kept_source_id",
          "reason"
        ],
        "title": "DuplicateGroup"
      },
...
      "GenerateAssetRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "source_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Id"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "source_id"
        ],
        "title": "GenerateAssetRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryAtom": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Source Ids"
          },
          "event": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Event"
          },
          "people": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "People",
            "default": []
          },
          "place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place"
          },
          "objects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Objects",
            "default": []
          },
          "moods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Moods",
            "default": []
          },
          "themes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Themes",
            "default": []
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "salience": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Salience"
          },
          "verbatim_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verbatim Text"
          },
          "must_preserve": {
            "type": "boolean",
            "title": "Must Preserve",
            "default": true
          },
          "subject_kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "object",
                  "scene"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Kind"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/Detail"
            },
            "type": "array",
            "title": "Details",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source_ids",
          "event",
          "salience"
        ],
        "title": "MemoryAtom"
      },
      "Moment": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "atom_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Atom Ids"
          },
          "source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Source Ids"
          },
          "image_source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Image Source Ids",
            "default": []
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "people": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "People",
            "default": []
          },
          "place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place"
          },
          "moods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Moods",
            "default": []
          },
          "salience": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Salience"
          },
          "time_range": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "string",
                    "format": "date-time"
                  },
                  {
                    "type": "string",
                    "format": "date-time"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Range"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "atom_ids",
          "source_ids",
          "summary",
          "salience"
        ],
        "title": "Moment"
      },
...
      "PhotoSeen": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "subject_kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "object",
                  "scene"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Kind"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels",
            "default": []
          },
          "people": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "People",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_id"
        ],
        "title": "PhotoSeen"
      },
...
      "ProviderInfo": {
        "properties": {
          "reasoning": {
            "type": "string",
            "enum": [
              "live",
              "fixture",
              "heuristic"
            ],
            "title": "Reasoning"
          },
          "reasoning_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Model"
          },
          "embeddings": {
            "type": "string",
            "enum": [
              "siglip2",
              "fallback"
            ],
            "title": "Embeddings"
          },
          "embedding_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model"
          },
          "cross_modal": {
            "type": "boolean",
            "title": "Cross Modal",
            "description": "True when text queries and images share one embedding space (SigLIP 2)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reasoning",
          "embeddings",
          "cross_modal"
        ],
        "title": "ProviderInfo"
      },
...
      "RoleCandidates": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "candidate_moment_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Candidate Moment Ids"
          },
          "scores": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Scores"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role",
          "intent",
          "query",
          "candidate_moment_ids",
          "scores"
        ],
        "title": "RoleCandidates"
      },
...
      "ScoredBundle": {
        "properties": {
          "seed": {
            "type": "string",
            "enum": [
              "anchor",
              "alternative",
              "baseline"
            ],
            "title": "Seed"
          },
          "moment_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Moment Ids"
          },
          "roles": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Roles",
            "description": "moment_id \u2192 narrative role"
          },
          "score": {
            "$ref": "#/components/schemas/BundleScore"
          },
          "expansion_log": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expansion Log",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "seed",
          "moment_ids",
          "roles",
          "score"
        ],
        "title": "ScoredBundle"
      },
...
      "StageTiming": {
        "properties": {
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "ms": {
            "type": "integer",
            "title": "Ms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "stage",
          "ms"
        ],
        "title": "StageTiming"
      },
      "StickerPickOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "src": {
            "type": "string",
            "title": "Src"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "chosen": {
            "type": "boolean",
            "title": "Chosen"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "matched": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched"
          },
          "color_fit": {
            "type": "number",
            "title": "Color Fit"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "src",
          "kind",
          "chosen",
          "score",
          "matched",
          "color_fit",
          "reason"
        ],
        "title": "StickerPickOut"
      },
      "StoryIntent": {
        "properties": {
          "intent": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Intent"
          },
          "title": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Title"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Keywords",
            "default": []
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleIntent"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 2,
            "title": "Roles",
            "description": "First role is always the Anchor."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "intent",
          "title",
          "roles"
        ],
        "title": "StoryIntent"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifierReport": {
        "properties": {
          "coverage": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Coverage"
          },
          "coherence": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Coherence"
          },
          "complementarity": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Complementarity"
          },
          "groundedness": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Groundedness"
          },
          "visual_diversity": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Visual Diversity"
          },
          "weakest_moment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weakest Moment Id"
          },
          "replacement_moment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replacement Moment Id"
          },
          "rationale": {
            "type": "string",
            "maxLength": 400,
            "title": "Rationale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "coverage",
          "coherence",
          "complementarity",
          "groundedness",
          "visual_diversity",
          "rationale"
        ],
        "title": "VerifierReport"
      },
...
      "WeaveRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "WeaveRequest"
      },
      "WeaveResult": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "0.1.0",
            "title": "Schema Version",
            "default": "0.1.0"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "providers": {
            "$ref": "#/components/schemas/ProviderInfo"
          },
          "atoms": {
            "items": {
              "$ref": "#/components/schemas/MemoryAtom"
            },
            "type": "array",
            "title": "Atoms"
          },
          "duplicate_groups": {
            "items": {
              "$ref": "#/components/schemas/DuplicateGroup"
            },
            "type": "array",
            "title": "Duplicate Groups"
          },
          "moments": {
            "items": {
              "$ref": "#/components/schemas/Moment"
            },
            "type": "array",
            "title": "Moments"
          },
          "story": {
            "$ref": "#/components/schemas/StoryIntent"
          },
          "role_candidates": {
            "items": {
              "$ref": "#/components/schemas/RoleCandidates"
            },
            "type": "array",
            "title": "Role Candidates"
          },
          "seeds": {
            "items": {
              "$ref": "#/components/schemas/ScoredBundle"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 1,
            "title": "Seeds"
          },
          "selected": {
            "$ref": "#/components/schemas/ScoredBundle"
          },
          "verifier": {
            "$ref": "#/components/schemas/VerifierReport"
          },
          "verifier_mode": {
            "type": "string",
            "enum": [
              "live",
              "fixture",
              "heuristic"
            ],
            "title": "Verifier Mode"
          },
          "replacement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replacement",
            "description": "\"old \u2192 new\" when the one allowed replacement was applied."
          },
          "baseline": {
            "$ref": "#/components/schemas/ScoredBundle"
          },
          "palette": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "minItems": 3,
            "title": "Palette"
          },
          "uses_timeline": {
            "type": "boolean",
            "title": "Uses Timeline"
          },
          "art": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtDirectionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "stickers": {
            "items": {
              "$ref": "#/components/schemas/StickerPickOut"
            },
            "type": "array",
            "title": "Stickers",
            "default": []
          },
          "day": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day",
            "description": "The day most photos were taken on; the memory is filed under it."
          },
          "timings": {
            "items": {
              "$ref": "#/components/schemas/StageTiming"
            },
            "type": "array",
            "title": "Timings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "created_at",
          "providers",
          "atoms",
          "duplicate_groups",
          "moments",
          "story",
          "role_candidates",
          "seeds",
          "selected",
          "verifier",
          "verifier_mode",
          "baseline",
          "palette",
          "uses_timeline",
          "timings"
        ],
        "title": "WeaveResult"
      }
    }
  }
}